LayoutTests/fast/canvas/canvas-too-large-to-draw-expected.txt

CONSOLE MESSAGE: line 33: Total canvas memory use exceeds the maximum limit (15 MB).
This test requires Internals.

 
